[ETCM-841] Rename CommonMessages to BaseETH6XMessages
