+++ src/benchmark/scala/io/iohk/ethereum/rlp/RLPSpeedSuite.scala
@@ -4,7 +4,7 @@ import akka.util.ByteString
 import io.iohk.ethereum.ObjectGenerators
 import io.iohk.ethereum.domain.Block._
 import io.iohk.ethereum.domain._
-import io.iohk.ethereum.network.p2p.messages.CommonMessages.SignedTransactions._
+import io.iohk.ethereum.network.p2p.messages.BaseETH6XMessages.SignedTransactions._
 import io.iohk.ethereum.utils.Logger
 import io.iohk.ethereum.utils.Hex
 import org.scalacheck.Gen
