+++ src/main/scala/io/iohk/ethereum/domain/Block.scala
@@ -29,7 +29,7 @@ case class Block(header: BlockHeader, body: BlockBody) {
 object Block {
 
   implicit class BlockEnc(val obj: Block) extends RLPSerializable {
-    import io.iohk.ethereum.network.p2p.messages.CommonMessages.SignedTransactions._
+    import io.iohk.ethereum.network.p2p.messages.BaseETH6XMessages.SignedTransactions._
 
     override def toRLPEncodable: RLPEncodeable = RLPList(
       obj.header.toRLPEncodable,
@@ -39,7 +39,7 @@ object Block {
   }
 
   implicit class BlockDec(val bytes: Array[Byte]) extends AnyVal {
-    import io.iohk.ethereum.network.p2p.messages.CommonMessages.SignedTransactions._
+    import io.iohk.ethereum.network.p2p.messages.BaseETH6XMessages.SignedTransactions._
     def toBlock: Block = rawDecode(bytes) match {
       case RLPList(header: RLPList, stx: RLPList, uncles: RLPList) =>
         Block(
