+++ src/main/scala/io/iohk/ethereum/domain/BlockBody.scala
@@ -32,7 +32,7 @@ object BlockBody {
 
   implicit class BlockBodyEnc(msg: BlockBody) extends RLPSerializable {
     override def toRLPEncodable: RLPEncodeable = {
-      import io.iohk.ethereum.network.p2p.messages.CommonMessages.SignedTransactions._
+      import io.iohk.ethereum.network.p2p.messages.BaseETH6XMessages.SignedTransactions._
 
       blockBodyToRlpEncodable(
         msg,
@@ -62,7 +62,7 @@ object BlockBody {
 
   implicit class BlockBodyRLPEncodableDec(val rlpEncodeable: RLPEncodeable) {
     def toBlockBody: BlockBody = {
-      import io.iohk.ethereum.network.p2p.messages.CommonMessages.SignedTransactions._
+      import io.iohk.ethereum.network.p2p.messages.BaseETH6XMessages.SignedTransactions._
 
       rlpEncodableToBlockBody(
         rlpEncodeable,
