+++ src/main/scala/io/iohk/ethereum/jsonrpc/RawTransactionCodec.scala
@@ -2,7 +2,7 @@ package io.iohk.ethereum.jsonrpc
 
 import akka.util.ByteString
 import io.iohk.ethereum.domain.SignedTransaction
-import io.iohk.ethereum.network.p2p.messages.CommonMessages.SignedTransactions.SignedTransactionEnc
+import io.iohk.ethereum.network.p2p.messages.BaseETH6XMessages.SignedTransactions.SignedTransactionEnc
 import io.iohk.ethereum.rlp
 
 object RawTransactionCodec {
