+++ src/main/scala/io/iohk/ethereum/network/handshaker/EtcNodeStatus63ExchangeState.scala
@@ -1,17 +1,17 @@
 package io.iohk.ethereum.network.handshaker
 
 import io.iohk.ethereum.network.EtcPeerManagerActor.{PeerInfo, RemoteStatus}
-import io.iohk.ethereum.network.p2p.messages.{CommonMessages, ProtocolVersions}
+import io.iohk.ethereum.network.p2p.messages.{BaseETH6XMessages, ProtocolVersions}
 import io.iohk.ethereum.network.p2p.{Message, MessageSerializable}
 
 case class EtcNodeStatus63ExchangeState(
     handshakerConfiguration: EtcHandshakerConfiguration
-) extends EtcNodeStatusExchangeState[CommonMessages.Status] {
+) extends EtcNodeStatusExchangeState[BaseETH6XMessages.Status] {
 
   import handshakerConfiguration._
 
   def applyResponseMessage: PartialFunction[Message, HandshakerState[PeerInfo]] = {
-    case status: CommonMessages.Status =>
+    case status: BaseETH6XMessages.Status =>
       applyRemoteStatusMessage(RemoteStatus(status))
   }
 
@@ -19,7 +19,7 @@ case class EtcNodeStatus63ExchangeState(
     val bestBlockHeader = getBestBlockHeader()
     val chainWeight = blockchain.getChainWeightByHash(bestBlockHeader.hash).get
 
-    val status = CommonMessages.Status(
+    val status = BaseETH6XMessages.Status(
       protocolVersion = ProtocolVersions.PV63.version,
       networkId = peerConfiguration.networkId,
       totalDifficulty = chainWeight.totalDifficulty,
