+++ src/test/scala/io/iohk/ethereum/ObjectGenerators.scala
@@ -11,7 +11,7 @@ import io.iohk.ethereum.domain.BlockHeader.HeaderExtraFields._
 import io.iohk.ethereum.domain._
 import io.iohk.ethereum.mpt.HexPrefix.bytesToNibbles
 import io.iohk.ethereum.mpt.{BranchNode, ExtensionNode, HashNode, LeafNode, MptNode, MptTraversals}
-import io.iohk.ethereum.network.p2p.messages.CommonMessages.NewBlock
+import io.iohk.ethereum.network.p2p.messages.BaseETH6XMessages.NewBlock
 import io.iohk.ethereum.network.p2p.messages.PV64
 import org.bouncycastle.crypto.AsymmetricCipherKeyPair
 import org.scalacheck.{Arbitrary, Gen, Shrink}
